matches page view

+++ server/routes/likes.js
@@ -79,16 +79,40 @@ likesRouter.delete('/:id', (req, res) => {
 		return res.status(401).send({ error: 'Invalid token or unauthorized' })
 	}
 
-	pool.query('DELETE FROM likes where id=?', req.params.id, (error, result) => {
+	pool.query('SELECT * FROM likes where id=?', req.params.id, (error, result) => {
 		if (result) {
-			pool.query(`UPDATE users SET fame = fame - 5 WHERE id = ${req.body.to}`)
-			res.status(200).end()
+			if (result[0].is_match === 1) {
+				console.log('Unmatching ids:', result[0].sender, result[0].receiver)
+				pool.query(`UPDATE likes SET is_match = 0 WHERE (sender=${result[0].receiver} AND receiver=${result[0].sender})`, (error, result) => {
+					if (result) {
+						pool.query('DELETE FROM likes where id=?', req.params.id, (error, result) => {
+							if (result) {
+								pool.query(`UPDATE users SET fame = fame - 5 WHERE id = ${req.body.to}`)
+								return res.status(200).end()
+							} else if (error) {
+								return res.status(500).send(error)
+							}
+						})
+					} else if (error) {
+						return res.status(500).send(error)
+					}
+				})
+			} else {
+				pool.query('DELETE FROM likes where id=?', req.params.id, (error, result) => {
+					if (result) {
+						pool.query(`UPDATE users SET fame = fame - 5 WHERE id = ${req.body.to}`)
+						return res.status(200).end()
+					} else if (error) {
+						return res.status(500).send(error)
+					}
+				})
+			}
 		} else if (error) {
-			res.status(500).send(error)
-		} else {
-			res.status(500).send({ error: 'Database error' })
+			return res.status(500).send(error)
 		}
 	})
+
+
 })
 
 module.exports = likesRouter
